Add fallback URL system for CDNs.
Russia is a shit and blocks a lot of western services. We want to be able to switch to an off-the-shelf CDN for engine downloads, but Russia gets in the way of that.

The fallback system allows 99% of users to use the main CDN, while the unfortunate ones that have Russia problems just go directly to our servers.